requirements.txt
PyQt6>=6.6.0
PyQt6-Fluent-Widgets>=1.5.0
Pillow>=10.0.0
psutil>=5.9.0
requests>=2.31.0
pyinstaller>=6.0.0
